Refactoring - renaming packages to original
