+++ app/src/test/java/com/tuenti/smsradar/SmsObserverTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.vaslabs.smsradar;
+package com.tuenti.smsradar;
 
 import org.junit.Before;
 import org.junit.Test;
