+++ smsradar-library/src/main/java/com/tuenti/smsradar/SmsStorage.java
@@ -13,21 +13,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.vaslabs.smsradar;
-
-import java.util.Date;
+package com.tuenti.smsradar;
 
 /**
- * Class created to work as time provider. This entity returns the current date. Avoid the client code to use System
- * .currentTimeMillis or create a new Date object directly.
+ * The SmsStorage has the responsibility to store the last sms intercepted by the library.
  *
  * @author Pedro Vicente Gómez Sánchez <pgomez@tuenti.com>
  * @author Manuel Peinado <mpeinado@tuenti.com>
  */
-class TimeProvider {
+interface SmsStorage {
 
-	public Date getDate() {
-		return new Date();
-	}
 
+	void updateLastSmsIntercepted(int smsId);
+
+	int getLastSmsIntercepted();
+
+	boolean isFirstSmsIntercepted();
 }
